+++ src/main/datetime.cpp
@@ -222,10 +222,10 @@ static double mktime00 (struct tm *tm)
     /* safety check for unbounded loops */
     if (year0 > 3000) {
 	excess = (int)(year0/2000) - 1;
-	year0 -= excess * 2000;
+	year0 -= excess * 2000;  /* FIXME arr 2007/06/01 */
     } else if (year0 < 0) {
 	excess = -1 - (int)(-year0/2000);
-	year0 -= excess * 2000;
+	year0 -= excess * 2000;  /* FIXME arr 2007/06/01 */
     }
 
     for(i = 0; i < tm->tm_mon; i++) day += days_in_month[i];
@@ -395,7 +395,7 @@ static struct tm * localtime0(const double *tp, const int local, struct tm *ltm)
 	res->tm_isdst = -1;
 
 	/* Try to fix up timezone differences */
-        diff = guess_offset(res)/60;
+        diff = int(guess_offset(res)/60);
 	shift = res->tm_min + 60*res->tm_hour;
 	res->tm_min -= diff;
 	validate_tm(res);
@@ -403,7 +403,7 @@ static struct tm * localtime0(const double *tp, const int local, struct tm *ltm)
 	/* now this might be a different day */
 	if(shift - diff < 0) res->tm_yday--;
 	if(shift - diff > 24) res->tm_yday++;	
-	diff2 = guess_offset(res)/60;
+	diff2 = int(guess_offset(res)/60);
 	if(diff2 != diff) {
 	    res->tm_min += (diff - diff2);
 	    validate_tm(res);
@@ -652,7 +652,7 @@ SEXP attribute_hidden do_asPOSIXct(SEXP call, SEXP op, SEXP args, SEXP env)
     PROTECT(ans = allocVector(REALSXP, n));
     for(i = 0; i < n; i++) {
 	double secs = REAL(VECTOR_ELT(x, 0))[i%nlen[0]], fsecs = floor(secs);
-	tm.tm_sec   = fsecs;
+	tm.tm_sec   = int(fsecs);
 	tm.tm_min   = INTEGER(VECTOR_ELT(x, 1))[i%nlen[1]];
 	tm.tm_hour  = INTEGER(VECTOR_ELT(x, 2))[i%nlen[2]];
 	tm.tm_mday  = INTEGER(VECTOR_ELT(x, 3))[i%nlen[3]];
@@ -713,7 +713,7 @@ SEXP attribute_hidden do_formatPOSIXlt(SEXP call, SEXP op, SEXP args, SEXP env)
     PROTECT(ans = allocVector(STRSXP, N));
     for(i = 0; i < N; i++) {
 	double secs = REAL(VECTOR_ELT(x, 0))[i%nlen[0]], fsecs = floor(secs);
-	tm.tm_sec   = fsecs;
+	tm.tm_sec   = int(fsecs);
 	tm.tm_min   = INTEGER(VECTOR_ELT(x, 1))[i%nlen[1]];
 	tm.tm_hour  = INTEGER(VECTOR_ELT(x, 2))[i%nlen[2]];
 	tm.tm_mday  = INTEGER(VECTOR_ELT(x, 3))[i%nlen[3]];
