+++ src/main/random.cpp
@@ -409,7 +409,7 @@ static void SampleReplace(int k, int n, int *y)
 {
     int i;
     for (i = 0; i < k; i++)
-	y[i] = n * unif_rand() + 1;
+	y[i] = int(n * unif_rand() + 1);
 }
 
 /* Equal probability sampling; without-replacement case */
@@ -420,7 +420,7 @@ static void SampleNoReplace(int k, int n, int *y, int *x)
     for (i = 0; i < n; i++)
 	x[i] = i;
     for (i = 0; i < k; i++) {
-	j = n * unif_rand();
+	j = int(n * unif_rand());
 	y[i] = x[j] + 1;
 	x[j] = x[--n];
     }
